added the qrScan response part of backend

+++ src/main/java/com/sih/craftshieldb/passport_service/Repo/CraftsRepo.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface CraftsRepo extends JpaRepository<Crafts, Integer> {
+public interface CraftsRepo extends JpaRepository<Crafts,Long> {
 }
